refactor for better design
