+++ src/client.rs
@@ -1,4 +1,5 @@
 mod body;
+mod ca;
 mod param;
 mod query;
 mod req;
@@ -7,10 +8,7 @@ pub mod resp;
 use std::{net::IpAddr, time::Duration};
 
 use ::serde::Deserialize;
-use magnus::{
-    Module, Object, RModule, RString, Ruby, TryConvert, Value, function, method, typed_data::Obj,
-    value::ReprValue,
-};
+use magnus::{Module, Object, RModule, Ruby, TryConvert, Value, function, method, typed_data::Obj};
 use wreq::Proxy;
 
 use crate::{
@@ -18,12 +16,13 @@ use crate::{
     client::{req::execute_request, resp::Response},
     cookie::Jar,
     emulate::Emulation,
-    error::{argument_error, option_value_error, wreq_error},
+    error::{option_value_error, wreq_error},
     extractor::Extractor,
     gvl,
     header::{Headers, OrigHeaders, UserAgent},
     http::Method,
     options::{NativeOption, Options},
+    utils::convert_path,
 };
 
 /// A builder for `Client`.
@@ -185,22 +184,14 @@ impl Builder {
             )
             .finish()?;
 
-        // Convert path-like objects (Pathname, etc.) for CA file options.
-        if let Some(value) = options.convert_present::<Value>("ca_file")?
-            && !value.is_nil()
-        {
-            builder.ca_file.set(Some(
-                convert_path(value).map_err(|e| option_value_error("ca_file", e))?,
-            ));
-        }
-        if let Some(value) = options.convert_present::<Value>("additional_ca_file")?
-            && !value.is_nil()
-        {
-            builder.additional_ca_file.set(Some(
-                convert_path(value).map_err(|e| option_value_error("additional_ca_file", e))?,
-            ));
-        }
-
+        extract_native_option!(
+            options, builder, ca_file,
+            Value =>? convert_path
+        );
+        extract_native_option!(
+            options, builder, additional_ca_file,
+            Value =>? convert_path
+        );
         extract_native_option!(
             options,
             builder,
@@ -217,16 +208,6 @@ impl Builder {
 
         Ok(builder)
     }
-}
-
-/// Convert a Ruby value to a file-system path `String`.
-///
-/// Accepts a plain `String` or any object responding to `to_path` (e.g. `Pathname`).
-fn convert_path(value: Value) -> Result<String, magnus::Error> {
-    if let Ok(path) = value.funcall::<_, _, RString>("to_path", ()) {
-        return path.to_string();
-    }
-    String::try_convert(value)
 }
 
 // ===== impl Client =====
@@ -273,38 +254,7 @@ impl Client {
             .map(|jar| jar.clone_store(ruby))
             .transpose()?;
 
-        // Resolve the CA option (at most one is set, enforced by reject_conflicts).
-        // Read file contents before releasing the GVL so I/O errors become ArgumentError.
-        // The bool indicates whether to augment system defaults (true) or replace them (false).
-        let ca_pem_data: Option<(Vec<u8>, bool)> = if let Some(path) = params.ca_file.take() {
-            let data = std::fs::read(&path)
-                .map_err(|e| argument_error(ruby, format!("ca_file: cannot read {path}: {e}")))?;
-            Some((data, false))
-        } else if let Some(pem) = params.ca_pem.take() {
-            Some((pem.into_bytes(), false))
-        } else if let Some(path) = params.additional_ca_file.take() {
-            let data = std::fs::read(&path).map_err(|e| {
-                argument_error(ruby, format!("additional_ca_file: cannot read {path}: {e}"))
-            })?;
-            Some((data, true))
-        } else {
-            params
-                .additional_ca_pem
-                .take()
-                .map(|pem| (pem.into_bytes(), true))
-        };
-
-        // Validate that PEM data contains at least one certificate.
-        if let Some((ref pem_bytes, _)) = ca_pem_data
-            && !pem_bytes
-                .windows(27)
-                .any(|w| w == b"-----BEGIN CERTIFICATE-----")
-        {
-            return Err(argument_error(
-                ruby,
-                "PEM data does not contain any certificates",
-            ));
-        }
+        let mut ca = ca::resolve(ruby, &mut params)?;
         let result = gvl::nogvl(|| {
             let mut builder = wreq::Client::builder();
 
@@ -436,14 +386,13 @@ impl Client {
             apply_option!(set_if_some, builder, params.tls_info, tls_info);
 
             // Custom CA certificate store.
-            if let Some((pem_bytes, augment)) = ca_pem_data {
-                let mut store_builder = wreq::tls::trust::CertStore::builder();
-                if augment {
-                    store_builder = store_builder.set_default_paths();
-                }
-                let store = store_builder.add_stack_pem_certs(&pem_bytes).build()?;
-                builder = builder.tls_cert_store(store);
-            }
+            apply_option!(
+                set_if_some_try_map,
+                builder,
+                ca,
+                tls_cert_store,
+                ca::into_cert_store
+            );
 
             // Network options.
             apply_option!(set_if_some, builder, params.proxy, proxy);
